feat(detected-fields): add `DetectedField` to `/models`
